# docs/swagger.yaml
basePath: /api
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Enterprise-grade webhook routing and data transformation platform
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Webhook Router API
  version: "1.0"
paths:
...
  /change-password:
    get:
      description: Serves the HTML change password page for users with default credentials
      produces:
      - text/html
      responses:
        "200":
          description: Change password page HTML
          schema:
            type: string
        "302":
          description: Redirect to login or admin page
          schema:
            type: string
        "404":
          description: Change password page not found
          schema:
            type: string
      summary: Serve change password page
      tags:
      - auth
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Updates user password and invalidates current session
      parameters:
      - description: New password
        in: formData
        name: password
        required: true
        type: string
      - description: Password confirmation
        in: formData
        name: confirm_password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "302":
          description: Redirect to login page with success message
          schema:
            type: string
        "400":
          description: Invalid password or passwords do not match
          schema:
            type: string
        "401":
          description: Invalid session
          schema:
            type: string
        "405":
          description: Method not allowed
          schema:
            type: string
        "500":
          description: Failed to change password
          schema:
            type: string
      summary: Process password change
      tags:
      - auth
...
  /login:
    get:
      description: Serves the HTML login page
      produces:
      - text/html
      responses:
        "200":
          description: Login page HTML
          schema:
            type: string
        "404":
          description: Login page not found
          schema:
            type: string
      summary: Serve login page
      tags:
      - auth
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Authenticates user credentials and creates a session
      parameters:
      - description: Username
        in: formData
        name: username
        required: true
        type: string
      - description: Password
        in: formData
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "302":
          description: Redirect to admin or change password page
          schema:
            type: string
        "401":
          description: Invalid credentials
          schema:
            type: string
        "405":
          description: Method not allowed
          schema:
            type: string
      summary: Process login
      tags:
      - auth
...
securityDefinitions:
  SessionAuth:
    in: cookie
    name: session
    type: apiKey
swagger: "2.0"